Refactors code to use dir for binary
- Split CLI code to `cmd` directory
- Move library code to its own module
